requirements.txt
alpaca-py
pandas
numpy
python-dotenv